Updates for macOS Catalina and crypto update for macOS
Various updates and fixes.
- Updated and fixed for macOS crypto (corecrypto)
- Added ID Viewer app sandbox
- Updated code to remove deprecations

+++ cardcomm/pkcs11/src/common/tlvbuffer.cpp
@@ -92,7 +92,7 @@ namespace eIDMW
 				}
 				//--- get Length
 //            int iNrBytes = ulLen - ulIndex - 1;     //calculate remaining bytes
-				int iNrBytes = ulLen - ulIndex;	//calculate remaining bytes
+				int iNrBytes = (int)(ulLen - ulIndex);	//calculate remaining bytes
 
 				if (!TlvDecodeLen(pucData + ulIndex, &iNrBytes, &ulFieldLen))
 				{
